Paste in most of the code to get character codes from XML file.
This code is from a prototype version that I made before deciding that this program would work and that it should be made properly as a new project file rather than using the prototype version.

The variable names need to be changed to make more sense, comments need to be added, and the boilerplate needs to be added yet.