Add statusbar and menubar with items as placeholders...
for later when stuff is put into them.

The menubar would have the File\Exit button close it, Tools\Options would open a window to change things like whether to search the XML file with pressing Enter/clicking a button or as you type as well as showing the version of the XML file and allowing the user to update it if there's a newer version available. Help\About would be the About window.